feat(stringCalculator) : refactoring(class 분리, 비즈니스 흐름 변경, ...)
