+++ src/test/java/racingcar/StringCalculatorTest.java
@@ -1,5 +1,6 @@
 package racingcar;
 
+import calculator.StringCalculator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -47,7 +48,8 @@ class StringCalculatorTest {
     }
 
     @DisplayName(value = "//와 \\n 문자 사이에 커스텀 구분자를 지정할 수 있다.")
-    @ParameterizedTest@ValueSource(strings = {"//;\n1;2;3"})
+    @ParameterizedTest
+    @ValueSource(strings = {"//;\n1;2;3"})
     void customDelimiter(final String text) {
         assertThat(calculator.add(text)).isSameAs(6);
     }
